added auth, comment controller, auth controller.
added components for myposts, show post, signup.
updated routes

+++ app/Http/Controllers/PostController.php
@@ -3,16 +3,27 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
+use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Resources\Post as PostResource;
+use Illuminate\Support\Facades\Auth;
 
 
 class PostController extends Controller
 {
+
     public function index()
     {
         //
         return PostResource::collection(Post::all());
+    }
+
+    public function myPosts()
+    {
+        //
+        $id = Auth::id();
+        $posts = Post::where('user_id', $id)->get();
+        return PostResource::collection($posts);
     }
 
     public function create()
@@ -42,7 +53,13 @@ class PostController extends Controller
 
     public function show($id)
     {
-        //
+        $post = Post::where('id',$id)
+            ->with(['comments' => function ($query) use ($id) {
+                $query->where('post_id', $id);
+            },'comments.user'])
+            ->first();
+
+        return new PostResource($post);
     }
 
     public function edit($id)
@@ -75,10 +92,17 @@ class PostController extends Controller
     {
         //
         $post = Post::findOrFail($id);
-        $post->delete();
+
+
+        if (Auth::id() == $post->user_id) {
+            $post->delete();
+            $message = 'Post successfully deleted';
+        } else {
+            $message = 'Unauthorized to delete post';
+        }
 
         return response()->json([
-            'message' => 'Post deleted!'
+            'message' => $message
         ]);
     }
 }
